requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.0.4
django-fsm==3.0.0
django-simple-history==3.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
filelock==3.18.0
identify==2.6.9
iniconfig==2.1.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
psycopg2==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
pytest-mock==3.14.0
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.30.0
